renamed bycurrency to by_currency
